backend/requirements.txt
fastapi[standard]
sqlmodel
psycopg[binary]
python-multipart
pyjwt
pwdlib[argon2]
prometheus-fastapi-instrumentator
